MVP on communication channels and message type
This is a large initial commit to get an MVP in working order in an effort
to flesh out the basic lines of communication between the daemon, CLI,
services, and user I/O. There is some unused code which may or may not be
here in the future and this is by no means linted or representative of what
contributing commits should look like. This is purely for MVP and quickly
getting up and running.