move toward a more flexible configuration format
Initially, this project was going to use viper as the configuration mgmt
system for the daemon, services, and backends. Unfortunately, we need a
bit more flexibility in the way we interact with our config and thus will
be moving toward using a structural representation of the config. This
commit is the first in that transition.